+++ app/views/administration/topics/index.html.slim
@@ -1,0 +1,26 @@
+- content_for(:title, 'All topics')
+- navigation_add('Dashboard', administration_root_path)
+
+.container
+  = render 'shared/notice'
+
+  .zero-bottom-space
+    .row
+      .col.s6
+        h5 All topics
+      .col.s6
+        h5.right= link_to 'Create new topic', new_administration_topic_path, class: 'link-as-button'
+  table.striped
+    thead
+      tr
+        th colspan="4"
+    tbody
+      - @topics.each do |topic|
+        tr
+          td= topic.name
+          td= link_to 'Show', administration_topic_path(topic), class: 'link-as-button'
+          td= link_to 'Edit', edit_administration_topic_path(topic), class: 'link-as-button'
+          td= link_to 'Destroy', administration_topic_path(topic), method: :delete, data: { confirm: 'Are you sure?' }, class: 'link-as-button'
+
+  .center
+    = paginate @topics
