+++ app/views/administration/topics/show.html.slim
@@ -1,0 +1,11 @@
+- content_for(:title, 'Show topic')
+- navigation_add('All topics', administration_topics_path)
+
+.container
+  = render 'shared/notice'
+
+  .zero-bottom-space
+    .row
+      h5= "Topic: #{@topic.name}"
+      .col.s6
+        h5= link_to 'Edit', edit_administration_topic_path(@topic), class: 'link-as-button'
