+++ app/views/layouts/_sidebar.html.slim
@@ -1,5 +1,4 @@
 ul#slide-out.sidenav
-  li= link_to "All topics", topics_path
   - if current_user
     - if policy(:dashboard).home?
       li= link_to "Admin Area", administration_root_path
