+++ app/views/layouts/administration/_mobile_menu.html.slim
@@ -1,4 +1,4 @@
-ul#slide-out.sidenav
+ul#slide-out.sidenav.highlighted
   li= main_nav_link('Dashboard', administration_root_path)
   li= main_nav_link('Users', administration_users_path)
-  li= main_nav_link("All topics", topics_path)
+  li= main_nav_link("Topics", administration_topics_path)
