--- app/views/topics/_form.html.slim
@@ -1,9 +1,0 @@
-center
-  .row.container.default-card.form-card.z-depth-1
-    = simple_form_for(topic) do |f|
-      .pink-text
-        = f.error_notification
-      .input_field.col.s6
-        = f.input :name, placeholder: "Topic name", label: false, autofocus: true
-      .field.col.s6
-        = f.button :submit, class: 'btn-zero-top-space'
