--- app/views/topics/edit.html.slim
@@ -1,5 +1,0 @@
-container
-  center
-    h5= "Edit '#{@topic.name}'"
-    = render 'form', topic: @topic
-    = link_to 'Back', topics_path, class: 'link-as-button-big'
