--- app/views/topics/new.html.slim
@@ -1,5 +1,0 @@
-container
-  center
-    h5 Create new topic
-    = render 'form', topic: @topic
-    = link_to 'Back', topics_path, class: 'link-as-button-big'
