--- app/views/topics/show.html.slim
@@ -1,11 +1,0 @@
-container
-  = render 'shared/notice'
-
-  .zero-bottom-space
-    .row
-      h5= "Topic: #{@topic.name}"
-      - if policy(:topic).edit?
-        .col.s6
-          h5= link_to 'Edit', edit_topic_path(@topic), class: 'link-as-button'
-      .col.s6
-        h5= link_to 'Back', topics_path, class: 'link-as-button'
