+++ spec/routing/administration/topics_routing_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe TopicsController, type: :routing do
+RSpec.describe Administration::TopicsController, type: :routing do
   describe 'routing' do
     it 'routes to #index' do
       expect(get: '/topics').to route_to('topics#index')
